refactor: Add PaginationBar component for displaying pagination
This commit adds a new PaginationBar component to the project. The PaginationBar component is responsible for displaying the current page number, total number of pages, and the number of items found. It also includes navigation links to the last and next pages if available. This component will be used in the SearchPage to enhance the pagination functionality of the application.